backend/requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask-Cors
Flask-Session
